moved node layer messages to separate package
