+++ node/message/path.go
@@ -15,28 +15,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with Alvalor.  If not, see <http://www.gnu.org/licenses/>.
 
-package node
+package message
 
-import (
-	"github.com/alvalor/alvalor-go/types"
-)
-
-// Status message shares our current best distance and locator hashes for our best path.
-type Status struct {
-	Distance uint64
-}
-
-// Sync message shares locator hashes from our current best path.
-type Sync struct {
-	Locators []types.Hash
-}
+import "github.com/alvalor/alvalor-go/types"
 
 // Path message shares a partial path from to our best header.
 type Path struct {
 	Headers []*types.Header
-}
-
-// Request is a download request for an inventory/transaction/block.
-type Request struct {
-	Hash types.Hash
 }
